feat: Add interface types
